requirements.txt
Django==5.2.2
djangorestframework==3.16.0
psycopg2==2.9.10
celery[redis]==5.5.3
flower==2.0.1